backend/requirements.txt
Django>=4.2,<5.0
gunicorn>=21.0
psycopg2-binary>=2.9
dj-database-url>=2.1
django-redis>=5.4
django-cors-headers>=4.3
boto3>=1.34
django-storages>=1.14
google-api-python-client>=2.143.0
google-auth>=2.35.0
google-auth-oauthlib>=1.2.0
channels>=4.1
channels-redis>=4.1
daphne>=4.1
transformers>=4.40
torch>=2.1
sentencepiece>=0.1.99
tiktoken>=0.5.0
wsaccel>=0.6.4
openai>=1.30.0
